fix(onboard): harden Ollama systemd follow-up
