+++ src/lib/onboard.ts
@@ -16,11 +16,11 @@ const {
   cliName,
   setOnboardBrandingAgent,
 }: typeof import("./onboard/branding") = require("./onboard/branding");
-const {
-  cleanupTempDir,
-  secureTempFile,
-}: typeof import("./onboard/temp-files") = require("./onboard/temp-files");
+const { cleanupTempDir }: typeof import("./onboard/temp-files") = require("./onboard/temp-files");
 const { stopStaleDashboardListenersForSandbox } = require("./onboard/stale-gateway-cleanup");
+const {
+  ensureOllamaLoopbackSystemdOverride,
+}: typeof import("./onboard/ollama-systemd") = require("./onboard/ollama-systemd");
 const {
   CUSTOM_BUILD_CONTEXT_WARN_BYTES,
   isInsideIgnoredCustomBuildContextPath,
@@ -168,8 +168,6 @@ const {
 
 const onboardProviders = require("./onboard/providers");
 const hermesProviderAuth = require("./hermes-provider-auth");
-
-const OLLAMA_SYSTEMD_OVERRIDE_PATH = "/etc/systemd/system/ollama.service.d/override.conf";
 
 type RemoteProviderConfigEntry = {
   label: string;
@@ -6645,108 +6643,6 @@ async function selectAndValidateOllamaModel(
   }
 }
 
-type OllamaLoopbackSystemdOverrideState = "not-applicable" | "ready" | "failed";
-
-function ensureOllamaLoopbackSystemdOverride(): OllamaLoopbackSystemdOverrideState {
-  if (process.platform !== "linux" || isWsl()) return "not-applicable";
-
-  const hasOllamaSystemdUnit = !!runCapture(
-    [
-      "sh",
-      "-c",
-      "command -v systemctl >/dev/null && [ -d /run/systemd/system ] && systemctl list-unit-files ollama.service --no-legend 2>/dev/null | head -n1",
-    ],
-    { ignoreError: true },
-  ).trim();
-  if (!hasOllamaSystemdUnit) return "not-applicable";
-
-  console.log("  Configuring Ollama systemd loopback override...");
-  console.log(
-    `  Applying an Ollama systemd override (OLLAMA_HOST=127.0.0.1:${OLLAMA_PORT}). ` +
-      "The next steps use sudo to write the drop-in, reload systemd, and restart the service; " +
-      "you may be prompted for your password.",
-  );
-  const sudoPrefix = isNonInteractive() || !process.stdin.isTTY ? "sudo -n" : "sudo";
-  const existingDropInResult = runShell(
-    `if [ -e ${shellQuote(OLLAMA_SYSTEMD_OVERRIDE_PATH)} ]; then ${sudoPrefix} cat ${shellQuote(OLLAMA_SYSTEMD_OVERRIDE_PATH)}; fi`,
-    { ignoreError: true, suppressOutput: true, timeout: 30_000 },
-  );
-  if (existingDropInResult.error || existingDropInResult.status !== 0) {
-    console.error("  Failed to inspect existing Ollama systemd override.");
-    console.error("  Refusing to continue because preserving existing Ollama settings is required.");
-    process.exit(1);
-  }
-  const existingDropIn = String(existingDropInResult.stdout || "");
-  const dropInBody = mergeOllamaLoopbackSystemdOverride(existingDropIn);
-  const tmpDropIn = secureTempFile("nemoclaw-ollama-override", ".conf");
-  let overrideFailed = false;
-  try {
-    fs.writeFileSync(tmpDropIn, dropInBody, { mode: 0o644 });
-    const overrideResult = runShell(
-      [
-        "set -e",
-        `pre_state=$(${sudoPrefix} systemctl show ollama --property=ActiveEnterTimestampMonotonic --property=MainPID --value 2>/dev/null | tr '\\n' ' ')`,
-        `${sudoPrefix} install -D -m 0644 ${shellQuote(tmpDropIn)} ${shellQuote(OLLAMA_SYSTEMD_OVERRIDE_PATH)}`,
-        `${sudoPrefix} systemctl daemon-reload`,
-        `${sudoPrefix} systemctl --no-block restart ollama`,
-        "for _ in $(seq 1 30); do",
-        `  current_state=$(${sudoPrefix} systemctl show ollama --property=ActiveEnterTimestampMonotonic --property=MainPID --value 2>/dev/null | tr '\\n' ' ')`,
-        `  if [ "$current_state" != "$pre_state" ] && ${sudoPrefix} systemctl is-active --quiet ollama; then exit 0; fi`,
-        "  sleep 1",
-        "done",
-        "exit 1",
-      ].join("\n"),
-      { ignoreError: true, timeout: 45_000 },
-    );
-    if (overrideResult.error || overrideResult.status !== 0) {
-      overrideFailed = true;
-    }
-  } finally {
-    cleanupTempDir(tmpDropIn, "nemoclaw-ollama-override");
-  }
-  if (overrideFailed) {
-    console.error("  Failed to apply Ollama systemd loopback override.");
-    console.error("  Refusing to continue with a potentially non-loopback Ollama bind.");
-    process.exit(1);
-  }
-
-  // The restart may briefly drop Ollama. Clear the cached successful probe so
-  // the readiness loop checks the daemon that systemd just restarted.
-  resetOllamaHostCache();
-  for (let i = 0; i < 10; i++) {
-    if (findReachableOllamaHost()) return "ready";
-    sleep(1);
-  }
-  return "failed";
-}
-
-function mergeOllamaLoopbackSystemdOverride(existingDropIn: string): string {
-  const desiredLine = `Environment="OLLAMA_HOST=127.0.0.1:${OLLAMA_PORT}"`;
-  const lines = existingDropIn.trimEnd().length > 0 ? existingDropIn.trimEnd().split(/\r?\n/) : [];
-  const serviceStart = lines.findIndex((line) => /^\s*\[Service\]\s*(?:[#;].*)?$/.test(line));
-  if (serviceStart === -1) {
-    return [...lines, ...(lines.length > 0 ? [""] : []), "[Service]", desiredLine].join("\n") + "\n";
-  }
-
-  let serviceEnd = lines.length;
-  for (let i = serviceStart + 1; i < lines.length; i++) {
-    if (/^\s*\[[^\]]+\]\s*(?:[#;].*)?$/.test(lines[i])) {
-      serviceEnd = i;
-      break;
-    }
-  }
-
-  const hostLines = lines
-    .slice(serviceStart + 1, serviceEnd)
-    .filter((line) => !/^\s*[#;]/.test(line) && /\bOLLAMA_HOST=/.test(line));
-  if (hostLines.at(-1)?.trim() === desiredLine) {
-    return lines.join("\n") + "\n";
-  }
-
-  lines.splice(serviceEnd, 0, desiredLine);
-  return lines.join("\n") + "\n";
-}
-
 async function setupNim(
   gpu: ReturnType<typeof nim.detectGpu>,
   sandboxName: string | null = null,
@@ -7649,7 +7545,7 @@ async function setupNim(
       } else if (selected.key === "ollama") {
         if (!checkOllamaPortsOrWarn()) continue selectionLoop;
         let ollamaReady = ollamaRunning;
-        const overrideState = ensureOllamaLoopbackSystemdOverride();
+        const overrideState = ensureOllamaLoopbackSystemdOverride({ isNonInteractive });
         if (overrideState === "ready") {
           ollamaReady = true;
         } else if (overrideState === "failed") {
@@ -7816,7 +7712,7 @@ async function setupNim(
           // daemon with our own `ollama serve`). This also repairs older
           // NemoClaw-created overrides that exposed raw Ollama on all interfaces.
           // WSL and non-systemd Linux fall back to a manual loopback launch.
-          const overrideState = ensureOllamaLoopbackSystemdOverride();
+          const overrideState = ensureOllamaLoopbackSystemdOverride({ isNonInteractive });
           if (overrideState === "failed") {
             console.error(
               "  Ollama systemd restart did not recover after applying the loopback override.",
